Do not set the surface ID on platform window before it is created
:Release Notes:
Do not set the surface ID on platform window before it is created.

:Detailed Notes:
After setting the PID on Chromium browser desktop window tree host
initialization, it would crash as the platform window would still
not exist. This change checks that condition, and stores the
surface id while the platform window is not created.

:Testing Performed:

:QA Notes:

:Issues Addressed:

Change-Id: Iff8bbae07578ca751266da0dd1ccea4acda6ce10
Reviewed-by: Jose Dapena <jose.dapena@lge.com>
Tested-by: Jose Dapena <jose.dapena@lge.com>

(cherry-picked from
https://github.com/webosose/chromium53/commit/407e64fc3560800d5cdac7ed50d1e21e720a43d8)